@@ -1,5 +1,6 @@
 package com.snowgears.shop.util;
 
+import com.snowgears.shop.Shop;
 import net.md_5.bungee.api.chat.ComponentBuilder;
 import net.md_5.bungee.api.chat.HoverEvent;
 import net.md_5.bungee.api.chat.TextComponent;
@@ -27,7 +28,7 @@ import java.net.URL;
 
 public class UpdateChecker {
 
-    private final JavaPlugin javaPlugin;
+    private final Shop plugin;
     private final String localPluginVersion;
     private String spigotPluginVersion;
 
@@ -41,9 +42,10 @@ public class UpdateChecker {
     private static final Permission UPDATE_PERM = new Permission("shop.update", PermissionDefault.FALSE);
     private static final long CHECK_INTERVAL = 12_000; //In ticks.
 
-    public UpdateChecker(final JavaPlugin javaPlugin) {
-        this.javaPlugin = javaPlugin;
-        this.localPluginVersion = javaPlugin.getDescription().getVersion();
+
+    public UpdateChecker(final Shop plugin) {
+        this.plugin = plugin;
+        this.localPluginVersion = plugin.getDescription().getVersion();
     }
 
     public void checkForUpdate() {
@@ -51,7 +53,7 @@ public class UpdateChecker {
             @Override
             public void run() {
                 //The request is executed asynchronously as to not block the main thread.
-                Bukkit.getScheduler().runTaskAsynchronously(javaPlugin, () -> {
+                plugin.getFoliaLib().getScheduler().runAsync(task -> {
                     //Request the current version of your plugin on SpigotMC.
                     try {
                         final HttpsURLConnection connection = (HttpsURLConnection) new URL("https://api.spigotmc.org/legacy/update.php?resource=" + ID).openConnection();
@@ -71,7 +73,7 @@ public class UpdateChecker {
                         Bukkit.getServer().getConsoleSender().sendMessage(ChatColor.translateAlternateColorCodes('&', embedVersions(UPDATE_MSG, localPluginVersion, spigotPluginVersion)));
 
                         //Register the PlayerJoinEvent
-                        Bukkit.getScheduler().runTask(javaPlugin, () -> Bukkit.getPluginManager().registerEvents(new Listener() {
+                        plugin.getFoliaLib().getScheduler().runNextTick(nextTask -> Bukkit.getPluginManager().registerEvents(new Listener() {
                             @EventHandler(priority = EventPriority.MONITOR)
                             public void onPlayerJoin(final PlayerJoinEvent event) {
                                 final Player player = event.getPlayer();
@@ -81,27 +83,27 @@ public class UpdateChecker {
                                 updateMsg.setClickEvent(new ClickEvent(ClickEvent.Action.OPEN_URL, "https://www.spigotmc.org/resources/shop-the-intuitive-shop-plugin.9628/updates"));
                                 player.spigot().sendMessage(updateMsg);
                             }
-                        }, javaPlugin));
+                        }, plugin));
                     }
                     /* RUNNING SHOP DEV VERSION */
                     if (compareVersions(localPluginVersion, spigotPluginVersion) > 0) {
                         Bukkit.getServer().getConsoleSender().sendMessage(ChatColor.translateAlternateColorCodes('&', embedVersions(DEV_VERSION_MSG, localPluginVersion, spigotPluginVersion)));
 
                         //Register the PlayerJoinEvent
-                        Bukkit.getScheduler().runTask(javaPlugin, () -> Bukkit.getPluginManager().registerEvents(new Listener() {
+                        plugin.getFoliaLib().getScheduler().runNextTick(nextTask -> Bukkit.getPluginManager().registerEvents(new Listener() {
                             @EventHandler(priority = EventPriority.MONITOR)
                             public void onPlayerJoin(final PlayerJoinEvent event) {
                                 final Player player = event.getPlayer();
                                 if (!player.isOp()) return;
                                 player.sendMessage(ChatColor.translateAlternateColorCodes('&', embedVersions(DEV_VERSION_MSG, localPluginVersion, spigotPluginVersion)));
                             }
-                        }, javaPlugin));
+                        }, plugin));
                     }
 
                     cancel(); //Cancel the runnable as an update has been found.
                 });
             }
-        }.runTaskTimer(javaPlugin, 0, CHECK_INTERVAL);
+        }.runTaskTimer(plugin, 0, CHECK_INTERVAL);
     }
 
     private String embedVersions(String msg, String runningVersion, String latestReleasedVersion) {
